requirements.txt
langchain_community 
tiktoken 
langchain-openai 
langchainhub 
chromadb 
langchain
python-dotenv

